+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # 4.0.0-alpha.4
 
 * Changed `Rant::run()` to return a `RantValue` and added `Rant::run_into_string()` to explicitly return a `String` output
+* Added support for whitespace normalization via `[whitespace-fmt]`.
 * Added new stdlib functions:
   * `[insert]`
   * `[remove]`
   * `[take]`
+  * `[whitespace-fmt]`
 
 # 4.0.0-alpha.3